refactor: transform lesson item form modals into pages
