+++ apps/web/app/modules/Admin/LessonItems/CreateNewQuestion.page.tsx
@@ -9,14 +9,6 @@ import { queryClient } from "~/api/queryClient";
 import { Button } from "~/components/ui/button";
 import { Checkbox } from "~/components/ui/checkbox";
 import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "~/components/ui/dialog";
-import {
   Form,
   FormControl,
   FormField,
@@ -32,8 +24,9 @@ import {
   SelectTrigger,
   SelectValue,
 } from "~/components/ui/select";
-import { useState } from "react";
 import Editor from "~/components/RichText/Editor";
+import { CreatePageHeader } from "~/modules/Admin/components";
+import { useNavigate } from "@remix-run/react";
 
 const questionFormSchema = z.object({
   questionType: z.enum([
@@ -49,37 +42,22 @@ const questionFormSchema = z.object({
   state: z.enum(["draft", "published"]),
   authorId: z.string().uuid("Invalid author ID."),
   solutionExplanation: z.string().optional(),
-});
-
-const answerOptionsSchema = z.object({
   options: z.array(
     z.object({
       value: z.string().min(1, "Option text is required"),
       isCorrect: z.boolean(),
       position: z.number(),
-    })
+    }),
   ),
 });
 
 type QuestionFormValues = z.infer<typeof questionFormSchema>;
-type AnswerOptionsFormValues = z.infer<typeof answerOptionsSchema>;
 
-export const CreateNewQuestion = ({
-  open,
-  onOpenChange,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}) => {
-  const [step, setStep] = useState<"question" | "options">("question");
-  const [createdQuestionId, setCreatedQuestionId] = useState<string | null>(
-    null
-  );
-
+export default function CreateNewQuestionPage() {
   const { mutateAsync: createQuestion } = useCreateQuestionItem();
   const { mutateAsync: assignAnswerOption } = useUpdateQuestionOptions();
   const { data: currentUser } = useCurrentUserSuspense();
-
+  const navigate = useNavigate();
   const questionForm = useForm<QuestionFormValues>({
     resolver: zodResolver(questionFormSchema),
     defaultValues: {
@@ -88,18 +66,18 @@ export const CreateNewQuestion = ({
       state: "draft",
       authorId: currentUser.id,
       solutionExplanation: "",
-    },
-  });
-
-  const answerOptionsForm = useForm<AnswerOptionsFormValues>({
-    resolver: zodResolver(answerOptionsSchema),
-    defaultValues: {
-      options: [],
+      options: [
+        {
+          value: "",
+          isCorrect: false,
+          position: 0,
+        },
+      ],
     },
   });
 
   const { fields, append, remove } = useFieldArray({
-    control: answerOptionsForm.control,
+    control: questionForm.control,
     name: "options",
   });
 
@@ -107,107 +85,35 @@ export const CreateNewQuestion = ({
     try {
       const response = await createQuestion({ data: values });
       const newQuestionId = response.data.questionId;
-      setCreatedQuestionId(newQuestionId);
+
+      if (newQuestionId) {
+        const options = values.options.map((option, index) => ({
+          questionId: newQuestionId,
+          optionText: option.value,
+          isCorrect: option.isCorrect,
+          position: index,
+        }));
+
+        await assignAnswerOption({
+          data: options,
+          questionId: newQuestionId,
+        });
+
+        await queryClient.invalidateQueries({
+          queryKey: ALL_LESSON_ITEMS_QUERY_KEY,
+        });
+
+        navigate(`/admin/lesson-items/${newQuestionId}`);
+        questionForm.reset();
+      }
 
       if (values.questionType === "open_answer") {
-        onOpenChange(false);
         queryClient.invalidateQueries({ queryKey: ALL_LESSON_ITEMS_QUERY_KEY });
-      } else {
-        setStep("options");
       }
     } catch (error) {
-      console.error("Error creating question:", error);
+      console.error("Error while creating question:", error);
     }
   };
-
-  const onOptionsSubmit = async (values: AnswerOptionsFormValues) => {
-    try {
-      if (!createdQuestionId) return;
-
-      const options = values.options.map((option, index) => ({
-        questionId: createdQuestionId,
-        optionText: option.value,
-        isCorrect: option.isCorrect,
-        position: index,
-      }));
-
-      await assignAnswerOption({
-        data: options,
-        questionId: createdQuestionId,
-      });
-
-      onOpenChange(false);
-      await queryClient.invalidateQueries({
-        queryKey: ALL_LESSON_ITEMS_QUERY_KEY,
-      });
-      setStep("question");
-      questionForm.reset();
-      answerOptionsForm.reset();
-    } catch (error) {
-      console.error("Error assigning answer options:", error);
-    }
-  };
-
-  const renderAnswerOptionsForm = () => (
-    <Form {...answerOptionsForm}>
-      <form
-        onSubmit={answerOptionsForm.handleSubmit(onOptionsSubmit)}
-        className="space-y-4"
-      >
-        <div className="space-y-4">
-          {fields.map((field, index) => (
-            <div key={field.id} className="flex items-center space-x-2">
-              <Controller
-                name={`options.${index}.value`}
-                control={answerOptionsForm.control}
-                render={({ field }) => (
-                  <Input {...field} placeholder="Enter answer option" />
-                )}
-              />
-              <Controller
-                name={`options.${index}.isCorrect`}
-                control={answerOptionsForm.control}
-                render={({ field }) => (
-                  <FormItem>
-                    <div className="flex items-center gap-2">
-                      <Label htmlFor={`isCorrect-${index}`}>Correct</Label>
-                      <Checkbox
-                        id={`isCorrect-${index}`}
-                        checked={field.value}
-                        onCheckedChange={(checked) => field.onChange(checked)}
-                      />
-                    </div>
-                  </FormItem>
-                )}
-              />
-              <Button
-                type="button"
-                variant="destructive"
-                onClick={() => remove(index)}
-              >
-                Remove
-              </Button>
-            </div>
-          ))}
-          <Button
-            type="button"
-            onClick={() =>
-              append({
-                value: "",
-                isCorrect: false,
-                position: fields.length,
-              })
-            }
-          >
-            Add Option
-          </Button>
-        </div>
-        <DialogFooter>
-          <Button type="submit">Create Options</Button>
-        </DialogFooter>
-      </form>
-    </Form>
-  );
 
   const renderQuestionForm = () => (
     <Form {...questionForm}>
@@ -304,28 +210,74 @@ export const CreateNewQuestion = ({
             </FormItem>
           )}
         />
-        <DialogFooter>
-          <Button type="submit">Next</Button>
-        </DialogFooter>
+
+        <Label className="space-y-4 flex flex-col">
+          <span>Answer Options</span>
+          {fields.map((field, index) => (
+            <div key={field.id} className="flex items-center space-x-2">
+              <Controller
+                name={`options.${index}.value`}
+                control={questionForm.control}
+                render={({ field }) => (
+                  <Input {...field} placeholder="Enter answer option" />
+                )}
+              />
+              <Controller
+                name={`options.${index}.isCorrect`}
+                control={questionForm.control}
+                render={({ field }) => (
+                  <FormItem>
+                    <div className="flex items-center gap-2">
+                      <Label htmlFor={`isCorrect-${index}`}>Correct</Label>
+                      <Checkbox
+                        id={`isCorrect-${index}`}
+                        checked={field.value}
+                        onCheckedChange={(checked) => field.onChange(checked)}
+                      />
+                    </div>
+                  </FormItem>
+                )}
+              />
+              <Button
+                type="button"
+                variant="destructive"
+                onClick={() => remove(index)}
+              >
+                Remove
+              </Button>
+            </div>
+          ))}
+          <Button
+            type="button"
+            className="w-min"
+            variant="outline"
+            onClick={() =>
+              append({
+                value: "",
+                isCorrect: false,
+                position: fields.length,
+              })
+            }
+          >
+            Add Option
+          </Button>
+        </Label>
+        <div className="flex w-full justify-end">
+          <Button type="submit" className="w-min">
+            Create question
+          </Button>
+        </div>
       </form>
     </Form>
   );
 
   return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="min-w-[750px]">
-        <DialogHeader>
-          <DialogTitle>
-            {step === "question" ? "Create New Question" : "Add Answer Options"}
-          </DialogTitle>
-          <DialogDescription>
-            {step === "question"
-              ? "Fill in the details to create a new question. Click next when you're done."
-              : "Add answer options for your question. Mark the correct answer(s)."}
-          </DialogDescription>
-        </DialogHeader>
-        {step === "question" ? renderQuestionForm() : renderAnswerOptionsForm()}
-      </DialogContent>
-    </Dialog>
+    <div className="flex flex-col gap-y-6">
+      <CreatePageHeader
+        title="Create New Question"
+        description="Fill in the details to create a new question. Click next when you're done."
+      />
+      {renderQuestionForm()}
+    </div>
   );
-};
+}
